build: add support for UEFI boot
UEFI boot can now be enabled via the '--uefi' flag.

We use the hyper UEFI binaries that ship with the release, but they can
optionally be specified manualy as well via the
'--hyper-uefi-binary-paths' command line option.

This makes use of image_utils for setting up the disk image as well as
retrieving boot firmware. The latter can optionaly be specified via the
'--uefi-firmware-path' parameter.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>