Introduce the idea of architecture execution mode
The idea here is that each architecture has a base name (e.g ARM, x86),
as well as an execution mode (such as x86_64, aarch64, riscv32, etc).
Prior to the introduction of this idea we would have to come up with
individual special-cased solutions to support multiple execution modes
of an architecture (see ULTRA_X86_POSTFIX and similar).

Hopefully with this change we will no longer need to do this, as this
idea of supporting multiple execution modes is now builtin into the
layout of the project. Each arch has an 'arch/basename' subdirectory
where it decides how to compile code based on the
ULTRA_ARCH_EXECUTION_MODE variable.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>