Rewrite the IO subsystem
This rewrite aims at reducing the amount of work that has to be done
per-arch in order to get working IO.

With this change, for the generic case an architecture only needs to
implement a few basic definitions in the arch/io_types.h header as well
as the io map/unmap helpers. It can also (optionally) implement
separate helpers for reading & writing port and memory IO if an
architecture needs separate handling for those.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>