kernel: introduce parameters & command line parsing helpers
This introdues the ability to define parameters that can be set via the
command line as well as sysfs (in the future).

The parameters are specified via simple parameter/early_parameter
macros, which automatically deduce the type and the respective ops to
use for converting the string specified by the command line to the
respective value, although manual macros like custom_parameter are also
available for full control.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>