common/format: clean up/modernize vsnprintf & related helpers
This bases the current vsnprintf implementation on the uACPI vsnprintf
implementation, which in turn is based on the Hyper vsnprintf
implementation. The previous implementation was messy and didn't handle
string precision correctly.

This also reworks the implementation to deal with struct string instead
of a C string, which greatly simplifies logic in consume{_one_of} etc.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>