log_ring: introduce a kernel log ring
This is a fully lockless & reentrant log ring that will be used as the
backbone of kernel logging, as well as syslog and /dev/kmsg. The
algorithm is not invented by me, but copied from Linux with my own
adjustments, completely reworked API and other enhancements. I've upstreamed
a fix for one of the issues I've noticed in the algorithm to Linux.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>